feat: melhora heurísticas de avaliação

+++ agent/tree.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 from game.utils import Player
 
 class Knot():    
-    def __init__(self, board: list[list[Player]], score: int, pos: tuple[int, int], depth: int):    
+    def __init__(self, board: list[list[Player]], score: float, pos: tuple[int, int], depth: int):    
         self.board = board
         self.score = score
         self.pos = pos
